Add files via upload
The purpose of this code is to evaluate traditional models in estimating the amount of carbon dioxide gas emissions caused by
fires are in the atmosphere. I used a relation between the volume of gas found in the atmosphere and the biomass in the jungle regions. Tools were satellite images as input for machine algorithms.